Fix Z hop in clicky unload macro
There is a comment indicating that an upwards Z hop was intended. It was instead calling the Z hop *down* macro, which caused the printhead to ram into the bed after QGL in my experience. This hop doesn't seem to be necessary at all, but we might as well correct it anyways.